Updated UiHelpers and UiBaseLocators
