+++ lib/helpers/UiHelpers.ts
@@ -1,12 +1,10 @@
-import {expect, Page} from "@playwright/test"
+import {Page} from "@playwright/test"
 import {StylesheetUiHelper} from "./StylesheetUiHelper";
 import {umbracoConfig} from "../../umbraco.config";
 import {PartialViewUiHelper} from "./PartialViewUiHelper";
 import {ScriptUiHelper} from "./ScriptUiHelper";
 import {TemplateUiHelper} from "./TemplateUiHelper";
 import {DictionaryUiHelper} from "./DictionaryUiHelper";
-import {UiBaseLocators} from "./UiBaseLocators";
-import {ConstantHelper} from "./ConstantHelper";
 
 export class UiHelpers {
   page: Page;
@@ -15,7 +13,6 @@ export class UiHelpers {
   dictionary: DictionaryUiHelper;
   script: ScriptUiHelper;
   template: TemplateUiHelper;
-  uiBaseLocators: UiBaseLocators;
 
   constructor(page: Page) {
     this.page = page;
@@ -23,29 +20,7 @@ export class UiHelpers {
     this.partialView = new PartialViewUiHelper(this.page);
     this.script = new ScriptUiHelper(this.page);
     this.template = new TemplateUiHelper(this.page);
-    this.dictionary = new DictionaryUiHelper(this.page)
-    this.uiBaseLocators = new UiBaseLocators(this.page);
-  }
-
-  async goToSettingsTreeItem(settingsTreeItemName: string) {
-    await this.uiBaseLocators.goToSection(ConstantHelper.sections.settings);
-    await this.page.getByLabel(settingsTreeItemName).click();
-  }
-
-  async clickDataElement(elementName: string, options: any = null) {
-    await this.page.click(`[data-element="${elementName}"]`, options);
-  }
-
-  async getDataElement(elementName: string) {
-    return this.page.locator(`[data-element="${elementName}"]`);
-  }
-
-  async isSuccessNotificationVisible() {
-    return await expect(this.page.locator('uui-toast-notification >> [color="positive"]')).toBeVisible();
-  }
-
-  async isErrorNotificationVisible() {
-    return await expect(this.page.locator('uui-toast-notification >> [color="danger"]')).toBeVisible();
+    this.dictionary = new DictionaryUiHelper(this.page);
   }
 
   async goToBackOffice() {
